+++ resources/views/budget/item/index.blade.php
@@ -19,84 +19,13 @@
             <div class="row">
                 <div class="col-12 col-lg-5 mx-auto p-2">
                     <div class="row budget-item g-3">
-                        <div class="col-12">
-                            <h2 class="display-6 mt-3 mb-3">
-                                <a class="btn btn-sm btn-primary text-end" href="">
-                                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil" viewBox="0 0 16 16">
-                                        <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0 1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"/>
-                                    </svg>
-                                </a>
-                                Budget Item
-                            </h2>
-                        </div>
-                        <div class="col-12">
-                            <div class="label">Name</div>
-                            {{ $item['name'] }}
-                        </div>
-                        <div class="col-6">
-                            <div class="label">Account</div>
-                            @if (array_key_exists($item['account'], $accounts))
-                                {{ $accounts[$item['account']]->name() }}
-                            @else
-                                Unknown Account (Please edit to fix)
-                            @endif
-                        </div>
-                        @if ($item['target_account'] !== null)
-                        <div class="col-6">
-                            <div class="label">Target Account</div>
-                            @if (array_key_exists($item['target_account'], $accounts))
-                                {{ $accounts[$item['target_account']]->name() }}
-                            @else
-                                Unknown Account (Please edit to fix)
-                            @endif
-                        </div>
-                        @endif
-                        <div class="col-12">
-                            <div class="label">Description</div>
-                            {{ $item['description'] ?? 'None set' }}
-                        </div>
-                        <div class="col-6">
-                            <div class="label">Start Date</div>
-                            {{ $item['start_date']->format('d/m/Y') }}
-                        </div>
-                        <div class="col-6">
-                            <div class="label">End Date</div>
-                            {{ ($item['end_date'] !== null) ? $item['end_date']->format('d/m/Y') : 'None set' }}
-                        </div>
-                        <div class="col-12">
-                            <div class="label">Amount & Type</div>
-                            <small><x-currency :currency="$item['currency']" /></small>{{ number_format($item['amount'], 2) }} <x-category :category="$item['category']" />
-                        </div>
-
-                        @if ($item['frequency']['type'] === 'monthly')
-                        <div class="col-12">
-                            <div class="label">Frequency</div>
-                            Monthly @if ($item['frequency']['day'] !== null) around the <x-ordinal :day="$item['frequency']['day']" /> of the month. @endif
-                        </div>
-                        @endif
-
-                        @if ($item['frequency']['type'] === 'annually')
-                        <div class="col-12">
-                            <div class="label">Frequency</div>
-                            Annually @if ($item['frequency']['day'] !== null) around the <x-ordinal :day="$item['frequency']['day']" /> of <x-month :month="$item['frequency']['month']" /> @else in <x-month :month="$item['frequency']['month']" /> @endif
-                        </div>
-                        @endif
-
-                        @if (
-                            $item['frequency']['type'] === 'monthly' &&
-                            array_key_exists('exclusions', $item)
-                        )
-                        <div class="col-12">
-                            <div class="label">Exclusions</div>
-                            Not required: {{ $item['exclusions'] }}
-                        </div>
-                        @endif
+                        <x-budget-item :accounts="$accounts" :item="$item" />
 
                         <div class="col-12">
                             <a class="btn btn-sm btn-outline-primary" href="{{ route('home') }}">
                                 Return
                             </a>
-                            <a class="btn btn-sm btn-dark" href="{{ route('budget.item.confirm-disable', ['item_id' => 'test']) }}">
+                            <a class="btn btn-sm btn-dark" href="{{ route('budget.item.confirm-disable', ['item_id' => $item['id']]) }}">
                                 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-moon" viewBox="0 0 16 16">
                                     <path d="M6 .278a.768.768 0 0 1 .08.858 7.208 7.208 0 0 0-.878 3.46c0 4.021 3.278 7.277 7.318 7.277.527 0 1.04-.055 1.533-.16a.787.787 0 0 1 .81.316.733.733 0 0 1-.031.893A8.349 8.349 0 0 1 8.344 16C3.734 16 0 12.286 0 7.71 0 4.266 2.114 1.312 5.124.06A.752.752 0 0 1 6 .278zM4.858 1.311A7.269 7.269 0 0 0 1.025 7.71c0 4.02 3.279 7.276 7.319 7.276a7.316 7.316 0 0 0 5.205-2.162c-.337.042-.68.063-1.029.063-4.61 0-8.343-3.714-8.343-8.29 0-1.167.242-2.278.681-3.286z"/>
                                 </svg>
@@ -123,8 +52,8 @@
                             :projection="$projection"
                             :hasAccounts="$has_accounts"
                             :hasBudget="$has_budget"/>
-                    </div>
                 </div>
+            </div>
 
             <x-requests />
